requirement.txt
streamlit==1.32.0
pdfplumber==0.10.3
PyMuPDF==1.24.3
rapidfuzz==3.6.1
sentence-transformers==2.6.1

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2

tiktoken==0.7.0
python-dotenv==1.0.1

google-generativeai==0.8.5
groq==0.6.0
